+++ src/components/layout/MobileSideMenu/MobileSideMenuContent.tsx
@@ -4,7 +4,7 @@ import styles from './css/MobileSideMenuContent.module.css';
 import useMobileSideMenu from '@/recoil/MobileSideMenu/useMobileSideMenu';
 import { MainCategory } from '@/model/category';
 import CategoryCard from '../../common/CategoryCard';
-import BackgroundOverlay from '@/components/common/BackgroundOverlay/BackgroundOverlay';
+import BackgroundOverlay from '@/components/common/BackgroundOverlay';
 import CloseButton from '@/components/common/CloseButton/CloseButton';
 
 type Props = {
